t9350: skip ISO-8859-? tests when the environment is always-UTF-8
In the BusyBox-w32 version that is currently under consideration for
MinGit for Windows (to reduce the .zip size, and to avoid problems with
the MSYS2 runtime), the UTF-16 environment present in Windows is
considered to be authoritative, and the 8-bit version is always in UTF-8
encoding.

As a consequence, the ISO-8859-? tests in t9350-fast-export (which try
to set GIT_AUTHOR_NAME to a ISO-8859-? encoded value) *must* fail in
that setup.

So let's detect when they would fail (due to an environment being purely
kept UTF-8 encoded), and skip those tests in that case.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>